Correction

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -12,8 +12,8 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t a;
-	
+	unsigned int a;
+
 	if (array != NULL && action != NULL)
 	{
 		for (a = 0; a < size; a++)
